Add files via upload
This is the Java implementation of the famous Knight's tour problem. This program simulates a chessboard and a single chess piece (the Knight). The Knight must visit every square on the board and cannot visit each square more than once. This implementation allows a user to enter the board size and starting position. In completing this project, I learned how to successfully implement Backtracking and Iterators in Java.